+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [v0.1.0]
+## v0.1.0
 
 - Downloaded Flickity v2.0.5, added libraries.yml file. Flickity config entity.
 
@@ -13,3 +13,5 @@ flickityInit. Improved options form.
 
 - Cacheability metadata. Move Flickity library to libraries/ folder. Fixed 
 adaptiveHeight option.
+
+- Description and README. Added CSS classes.